refactor(backend): extract architect crud module
